+++ search-service/src/main/java/com/gitlab/model/User.java
@@ -1,0 +1,57 @@
+package com.gitlab.model;
+
+import com.gitlab.enums.EntityStatus;
+import com.gitlab.enums.Gender;
+import lombok.*;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@EqualsAndHashCode(exclude = {"passport", "bankCardsSet", "shippingAddressSet"})
+@ToString
+@Table(name = "users", schema = "public", catalog = "postgres")
+public class User {
+
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long id;
+
+    @Column(unique = true, name = "email")
+    private String email;
+
+    @Column(unique = true, name = "username")
+    private String username;
+
+    @Column(name = "password")
+    private String password;
+
+    @Column(name = "security_question")
+    private String securityQuestion;
+
+    @Column(name = "answer_question")
+    private String answerQuestion;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "birth_date")
+    private LocalDate birthDate;
+
+    @Column(name = "gender")
+    @Enumerated(EnumType.STRING)
+    private Gender gender;
+
+    @Column(name = "phone_number")
+    private String phoneNumber;
+
+    @Column(name = "entity_status")
+    @Enumerated(EnumType.STRING)
+    private EntityStatus entityStatus;
+}