+++ search-service/src/main/java/com/gitlab/repository/ProductImageRepository.java
@@ -1,0 +1,9 @@
+package com.gitlab.repository;
+
+import com.gitlab.model.ProductImage;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductImageRepository extends ReadOnlyRepository<ProductImage, Long> {
+
+}