+++ search-service/src/main/java/com/gitlab/repository/ProductRepository.java
@@ -1,0 +1,10 @@
+package com.gitlab.repository;
+
+import com.gitlab.model.Product;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends ReadOnlyRepository<Product, Long> {
+
+    Iterable<Product> findByNameContainingIgnoreCase(String name);
+}